+++ apps/server/index.js
@@ -11,7 +11,7 @@
  */
 
 import { createApp } from "./app.js";
-import { env } from "./config/index.js";
+import { env } from "./config/env.js";
 import { logger } from "./utils/logger.js";
 
 const app = createApp();